src/azure-communication-services-integrate-voice-live-api/python/requirements.txt
# Azure Communication Services with Voice Live API - Python Implementation
# Core web framework and WebSocket support
fastapi>=0.115.2,<1.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# Azure SDK packages for Communication Services
azure-communication-callautomation>=1.4.0
azure-identity>=1.17.0
azure-eventgrid>=4.18.0

# HTTP client and JSON handling
httpx>=0.25.2
aiohttp>=3.9.1
pydantic>=2.8.2

# Configuration and environment management
python-dotenv>=1.0.1
pydantic-settings>=2.5.0

# Audio processing libraries
numpy>=1.24.0
scipy>=1.11.0

# Logging and monitoring
structlog==23.2.0

# Audio processing utilities
numpy==1.24.3

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
